# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ErTraining",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/editCourse/{id}": {
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor - modifica i dati del corso, si possono modificare nome, stato e descrizione",
                "description": "Edit course",
                "operationId": "editCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "state",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "state"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/getProgress/{id}": {
            "get": {
                "tags": [
                    "Progress"
                ],
                "summary": "Insegnate - Ritorna tutti i progress degli utenti che partecipano al corso indicato",
                "description": "Get all the users and their progress of a course",
                "operationId": "getProgress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/setStateProgress/{id}": {
            "put": {
                "tags": [
                    "Progress"
                ],
                "summary": "editCourse",
                "description": "Insegnate - Cambia lo stato da non abilitato ad abilitato",
                "operationId": "setStateProgress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "progress Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "state"
                                ],
                                "properties": {
                                    "state": {
                                        "type": "string",
                                        "format": "state"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/changeStateProgress/{id}": {
            "put": {
                "tags": [
                    "Progress"
                ],
                "summary": "editCourse",
                "description": "Studente - cambia lo stato da abilitato ad in corso, da in corso a finito e crea il prossimo progress",
                "operationId": "changeStateProgress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "progress Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "state"
                                ],
                                "properties": {
                                    "state": {
                                        "type": "string",
                                        "format": "state"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}